Revert "utils folder renamed as it was leading to a false warning"
This reverts commit 81379335055966e58d13d4555544a5976dd95fe6.
