+++ response-items/src/discrete-scale/utils/radio-params.js
