+++ response-items/src/ellipse-scale/ResponseItem.stories.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import ResponseItem from "./ResponseItem";
 import Icon from "./Icon";
